food_back/backend/requirements.txt
accelerate==1.11.0
anyio==4.11.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.60.1
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.4
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
polars==1.35.2
polars-runtime-32==1.35.2
psutil==7.1.3
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
scipy==1.16.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
timm==1.0.22
tokenizers==0.22.1
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers @ git+https://github.com/huggingface/transformers@266d3b05684ed5968c613f00b208909f03c72c0e
typer-slim==0.20.0
typing_extensions==4.15.0
ultralytics==8.3.228
ultralytics-thop==2.0.18
urllib3==2.5.0
Werkzeug==3.1.3
